Create initial fork for Ansible apt library
Add framework to solve objective:

Remove the library dependency that blocks Ansible's use of apt in a Virtual
env.

Keep the same functionality without using this library. The command line is
already being used, can the same behavior also be achieved without the library.

Root cause for adding the library was to fix this issue in 2013:
https://github.com/ansible/ansible/issues/3421