Reset refactoring to a less regressive level
The first attempt to refactor was intended to remove the library components,
and then "fill in the holes" with command line equivalents.

However, as the structure of this file is long functions with nested try/except
statements, it's very difficult to get the clean "holes" to fill-in while
ensuring functionality has not yet changed.

The original file was reverted and updated with less aggressive changes (for
this commit).

For example, instead of removing *package_status*, we will need to better
understand it's funcionality.

We need to determine if:
  * A package is already installed (so we can update *changed* appropriately)
  * Tthe version number (if given) is already installed
  * The package is upgradable
  * The package has files.