Create system_setup.sh for Vagrant system python debugging
If we are going to debug the behavior that we see that python-apt is currently
using, then we need to get it in a working functional state.

The demonstrations have thus far been about what hasn't worked (in a Python
virtualenv). However, now we'll need to either be system python
(setup_system.sh) or python in a virtualenv (setup_venv.sh). This is set in the
Vagrant file:

  config.vm.provision "shell" do |p|
    p.inline = "/vagrant/setup_system.sh"
  end