Convert "dpkg -l" command line to "dpkg-query -W -f=.."
A peer review from Stefan Taranu exposed that there is a more robust way to
collect the same information. Instead of parsing command line output, we can
gather the information directly dpkg-query in JSON output.

> I studied your last commit on the develop branch and I would replace the dpkg
> command with dpkg-query. The reason for picking this tool over dpkg is
> because it provides a formatting mode (--showformat= /-f) and you can produce
> a custom JSON format that can be parsed with json.loads().
>
> The fields that you can output is quite extensive, covering package status,
> package wanted status, version, etc.
>
> Here's an example:
>
> dpkg-query -W -f='{"package":"${binary:Package}", "version": "${Version}",
> "maintaner": "${Maintainer}", "status": "${db:Status-Abbrev}", "depends":
> ["${Depends}"]}' wget

This does rely on json.loads() to behave properly (given that some input may be
Unicode data). We will be supporting Python2 and Python3 versions of this even
though Python2 is technically already deprecated. Our reason for doing this is
to help people transition from Python2 to Python3.

If we do find any json.loads() issue, we can recover the code deleted by this
commit and make it a fall-back on any json.loads() issue.