Assume all packages have files until Issue is fixed
The reason that python_apt was originally used was to find if a package
has files (there are some that do not). In working through these cases,
I seem to have tripped on this original bug.

Stefan Taranu originally gave a suggestion that may work in this case
(although it was a dpkg command and the package is not installed). Glen
can help find that information.

In short, we are going to bypass this functionality for now as we may
just remove that functionality for simplicity. We are recording this
ticket to track for now and will return when we have a real-live test
case (so we can revisit our options/needs).

We are simply stubbing this out for now and assuming all packages have
files until this is better addressed (I have not yet seen a test case
where we haven't had files in packages)